Update aws-cs-fargate: remove usage of deprecated constructs from elastic load balancing and use app load balancing instead
